add some parsing for gre headers.
we only look for the GRE key subheader if it's a GRE version 0
packet, and only if the packet has a cksum or seq no header set.
any extra flags set means parsing i'm not in the mood to write
right now.

if this becomes important i can look at the tcpdump code for
inspiration.