replace memcmp in flow_cmp with a custom comparison with long words.
memcmp on openbsd looks completely unoptimised, so it ends up doing
a byte by byte comparison. comparing whole longs is a lot faster,
and can take shortcuts since we can align the memory we're comparing
and don't care really care about hte order of comparisons or which
exact byte is different. clang also seems to unroll the loop.

this appears to halve the cost of processing a packet, which is
pretty nuts.