prepare for storing flows captured on enc(4)
the idea is to capture esp tunnel mode traffic and store the outer
tunnel address and spi, and then use the existing packet parser for
the traffic inside the tunel.

this will give us a better chance to tie who is doing what on the
network. we should be able to tie a user to an spi with the iked
logs.